requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
django-crontab==0.7.1
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
dnspython==2.8.0
gunicorn==24.1.0
idna==3.10
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.10.1
pymongo==4.16.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
